+++ user/threads.h
@@ -3,10 +3,10 @@
 
 #include "sys.h"
 
-int thread_create(void (*start_routine)(void*), void *args);
-int thread_exit(long exitCode);
-int thread_join(long id);
-int thread_yield();
+long thread_create(void (*start_routine)(void*), void *args);
+long thread_exit(long exitCode);
+long thread_join(long id);
+long thread_yield();
 
 #endif
